+++ Source/ModularOnline/Private/Core/ModularOnlineNetSerializers.h
@@ -1,0 +1,10 @@
+// Copyright PoFig Games Studio. All Rights Reserved.
+
+#pragma once
+
+#include "Iris/Serialization/NetSerializer.h"
+
+namespace PoFigGames::Online::Private
+{
+	UE_NET_DECLARE_SERIALIZER(FModularUniqueNetIdNetSerializer, );
+}